refactor(dashboard): split api.ts (935 LOC) into 4 domain modules
api.ts monolith → barrel re-export + domain modules:
- api/core.ts (280): HTTP infrastructure, auth, retry, operator actions
- api/mcp.ts (83): MCP-over-HTTP client, autoresearch functions
- api/keeper.ts (187): keeper messaging (direct, operator-routed, SSE streaming)
- api/dashboard.ts (408): dashboard projections, resource fetchers, tool metrics

Circular imports fixed: api/board.ts, api/trpg.ts, api/actions.ts now import
from ./core and ./mcp instead of ../api barrel.

All existing imports from './api' continue to work via barrel re-export.
