+++ dashboard/src/api/core.ts
@@ -1,0 +1,280 @@
+// MASC Dashboard — HTTP infrastructure (auth, fetch, retry, operator actions)
+
+import type {
+  OperatorActionRequest,
+  OperatorActionResult,
+} from '../types'
+
+// --- Auth ---
+
+function getQueryParams(): URLSearchParams {
+  return new URLSearchParams(window.location.search)
+}
+
+export const DASHBOARD_AGENT_NAME_KEY = 'masc_dashboard_agent_name'
+
+function readStoredAgentName(): string | null {
+  try {
+    return localStorage.getItem(DASHBOARD_AGENT_NAME_KEY)?.trim() || null
+  } catch {
+    return null
+  }
+}
+
+export function currentDashboardActor(): string {
+  const params = getQueryParams()
+  return (
+    params.get('agent')?.trim()
+    || params.get('agent_name')?.trim()
+    || readStoredAgentName()
+    || 'dashboard'
+  )
+}
+
+function authHeaders(): Record<string, string> {
+  const params = getQueryParams()
+  const headers: Record<string, string> = {}
+  const token = params.get('token')
+  const storedAgent = readStoredAgentName()
+  const agent = params.get('agent') ?? params.get('agent_name') ?? storedAgent
+  if (token) headers['Authorization'] = `Bearer ${token}`
+  if (agent) headers['X-MASC-Agent'] = agent
+  return headers
+}
+
+export function jsonHeaders(): Record<string, string> {
+  return {
+    ...authHeaders(),
+    'Content-Type': 'application/json',
+  }
+}
+
+export const DEFAULT_GET_TIMEOUT_MS = 15_000
+export const DEFAULT_POST_TIMEOUT_MS = 30_000
+export const DEFAULT_MCP_TIMEOUT_MS = 60_000
+export const ROOM_TRUTH_GET_TIMEOUT_MS = 30_000
+const RETRYABLE_STATUS_CODES = new Set([408, 425, 429, 500, 502, 503, 504])
+
+export class ApiRequestError extends Error {
+  method: string
+  path: string
+  status?: number
+  statusText?: string
+  timeout: boolean
+
+  constructor(opts: {
+    method: string
+    path: string
+    status?: number
+    statusText?: string
+    timeout?: boolean
+    timeoutMs?: number
+  }) {
+    const method = opts.method.toUpperCase()
+    const timeout = opts.timeout === true
+    const message = timeout
+      ? `${method} ${opts.path}: timeout after ${opts.timeoutMs ?? 0}ms`
+      : `${method} ${opts.path}: ${opts.status ?? 'unknown'} ${opts.statusText ?? ''}`.trim()
+    super(message)
+    this.name = 'ApiRequestError'
+    this.method = method
+    this.path = opts.path
+    this.status = opts.status
+    this.statusText = opts.statusText
+    this.timeout = timeout
+  }
+}
+
+export async function fetchWithTimeout(path: string, init: RequestInit, timeoutMs: number): Promise<Response> {
+  const controller = new AbortController()
+  const timer = setTimeout(() => controller.abort(), timeoutMs)
+
+  try {
+    return await fetch(path, {
+      ...init,
+      signal: controller.signal,
+    })
+  } catch (err) {
+    if (err instanceof Error && err.name === 'AbortError') {
+      const method = typeof init.method === 'string' ? init.method.toUpperCase() : 'GET'
+      throw new ApiRequestError({
+        method,
+        path,
+        timeout: true,
+        timeoutMs,
+      })
+    }
+    throw err
+  } finally {
+    clearTimeout(timer)
+  }
+}
+
+export function defaultBoardVoter(): string {
+  const params = getQueryParams()
+  return (
+    params.get('agent')?.trim() ||
+    params.get('agent_name')?.trim() ||
+    'dashboard-user'
+  )
+}
+
+// --- Generic fetcher ---
+
+export type GetOptions = {
+  timeoutMs?: number
+}
+
+export async function get<T>(path: string, opts: GetOptions = {}): Promise<T> {
+  const res = await fetchWithTimeout(
+    path,
+    { headers: authHeaders() },
+    opts.timeoutMs ?? DEFAULT_GET_TIMEOUT_MS,
+  )
+  if (!res.ok) {
+    throw new ApiRequestError({
+      method: 'GET',
+      path,
+      status: res.status,
+      statusText: res.statusText,
+    })
+  }
+  return res.json() as Promise<T>
+}
+
+export function sleep(ms: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, ms))
+}
+
+function parseStatusFromMessage(message: string): number | null {
+  const match = message.match(/\b(\d{3})\b/)
+  if (!match) return null
+  const statusToken = match[1]
+  if (!statusToken) return null
+  const status = Number.parseInt(statusToken, 10)
+  return Number.isFinite(status) ? status : null
+}
+
+function isRetryableError(err: unknown): boolean {
+  if (err instanceof ApiRequestError) {
+    return err.timeout || (typeof err.status === 'number' && RETRYABLE_STATUS_CODES.has(err.status))
+  }
+
+  if (!(err instanceof Error)) return false
+  if (/timeout after \d+ms/i.test(err.message)) return true
+  const parsedStatus = parseStatusFromMessage(err.message)
+  return parsedStatus !== null && RETRYABLE_STATUS_CODES.has(parsedStatus)
+}
+
+export async function withRetries<T>(
+  operation: string,
+  run: () => Promise<T>,
+  retries = 2,
+): Promise<T> {
+  let attempt = 0
+
+  while (true) {
+    try {
+      return await run()
+    } catch (err) {
+      if (!isRetryableError(err) || attempt >= retries) throw err
+      const delayMs = 250 * (attempt + 1)
+      console.warn(`[dashboard/api] ${operation} failed (attempt ${attempt + 1}), retrying in ${delayMs}ms`, err)
+      await sleep(delayMs)
+      attempt += 1
+    }
+  }
+}
+
+export async function post<T>(
+  path: string,
+  body: unknown,
+  extraHeaders?: Record<string, string>,
+  timeoutMs = DEFAULT_POST_TIMEOUT_MS,
+): Promise<T> {
+  const res = await fetchWithTimeout(path, {
+    method: 'POST',
+    headers: {
+      ...jsonHeaders(),
+      ...(extraHeaders ?? {}),
+    },
+    body: JSON.stringify(body),
+  }, timeoutMs)
+  if (!res.ok) {
+    throw new ApiRequestError({
+      method: 'POST',
+      path,
+      status: res.status,
+      statusText: res.statusText,
+    })
+  }
+  return res.json() as Promise<T>
+}
+
+export async function postRaw(
+  path: string,
+  body: unknown,
+  extraHeaders?: Record<string, string>,
+  timeoutMs = DEFAULT_POST_TIMEOUT_MS,
+): Promise<string> {
+  const res = await fetchWithTimeout(path, {
+    method: 'POST',
+    headers: {
+      ...jsonHeaders(),
+      ...(extraHeaders ?? {}),
+    },
+    body: JSON.stringify(body),
+  }, timeoutMs)
+  if (!res.ok) {
+    throw new ApiRequestError({
+      method: 'POST',
+      path,
+      status: res.status,
+      statusText: res.statusText,
+    })
+  }
+  return res.text()
+}
+
+// --- Operator actions ---
+
+function operatorActionTimeoutMs(body: OperatorActionRequest): number {
+  switch (body.action_type) {
+    case 'keeper_message':
+    case 'keeper_recover':
+      return 90_000
+    case 'swarm_run_continue':
+      return 60_000
+    case 'swarm_run_rerun':
+      return 120_000
+    case 'swarm_run_abandon':
+      return 30_000
+    case 'social_sweep':
+    case 'lodge_tick':
+      return 45_000
+    default:
+      return DEFAULT_POST_TIMEOUT_MS
+  }
+}
+export function runOperatorAction(body: OperatorActionRequest): Promise<OperatorActionResult> {
+  return post('/api/v1/operator/action', body, undefined, operatorActionTimeoutMs(body))
+}
+
+export function confirmOperatorAction(
+  actor: string,
+  confirmToken: string,
+  decision: 'confirm' | 'deny' = 'confirm',
+): Promise<OperatorActionResult> {
+  return post('/api/v1/operator/confirm', {
+    actor,
+    confirm_token: confirmToken,
+    decision,
+  })
+}
+
+export function runCommandPlaneAction(
+  path: string,
+  body: Record<string, unknown>,
+): Promise<Record<string, unknown>> {
+  return post(path, body)
+}
